MoveisApp#5 miniRepair

+++ src/services/delete-rating/fetch-delete-rating.js
@@ -1,21 +1,21 @@
 import PropTypes from 'prop-types';
 
 async function DeleteRating(sessionId, filmId) {
-  const options = {
-    method: 'DELETE',
-    headers: {
-      accept: 'application/json',
-      'Content-Type': 'application/json;charset=utf-8',
-    },
-  };
+  if (sessionId) {
+    const url = `https://api.themoviedb.org/3/movie/${filmId}/rating?api_key=ac9f5a2225820c509d5d3cd533371ddb&guest_session_id=${sessionId}`;
 
-  const deleteRatingResult = await fetch(
-    `https://api.themoviedb.org/3/movie/${filmId}/rating?api_key=ac9f5a2225820c509d5d3cd533371ddb&guest_session_id=${sessionId}`,
-    options
-  )
-    .then((response) => response.json())
-    .then((response) => response.status_message);
-  return deleteRatingResult;
+    const options = {
+      method: 'DELETE',
+      headers: {
+        accept: 'application/json',
+        'Content-Type': 'application/json;charset=utf-8',
+      },
+    };
+
+    const deleteRatingResult = fetch(url, options).then((response) => response.json());
+    return deleteRatingResult;
+  }
+  return null;
 }
 
 export default DeleteRating;
